C++ tests work, pyFastBDT always returns 0

+++ include/FastBDT_C_API.h
@@ -51,9 +51,9 @@ extern "C" {
 
     void Load(void* ptr, char *weightfile);
 
-    void Predict(void *ptr, float *array, double *result);
+    void Predict(void *ptr, float *array, float *result);
 
-    // void PredictArray(void *ptr, float *array, double *result, unsigned int nEvents);
+    // void PredictArray(void *ptr, float *array, float *result, unsigned int nEvents);
 
     void Save(void* ptr, char *weightfile);
